Migrate crs3c to abstract class
Should reduce code duplication substantially, allows ConcurrentLruMap to be extracted, and prevents duplicate calls being made for headBucket and listObjects.
