+++ spring-cloud-aws-autoconfigure/src/main/java/io/awspring/cloud/autoconfigure/s3/S3AutoConfiguration.java
@@ -30,6 +30,7 @@ import io.awspring.cloud.s3.S3OutputStreamProvider;
 import io.awspring.cloud.s3.S3ProtocolResolver;
 import io.awspring.cloud.s3.S3Template;
 import io.awspring.cloud.s3.crossregion.CrossRegionS3Client;
+import io.awspring.cloud.s3.crossregion.DefaultCrossRegionS3Client;
 import java.util.Optional;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
@@ -102,7 +103,7 @@ public class S3AutoConfiguration {
 		@Bean
 		@ConditionalOnMissingBean
 		S3Client s3Client(S3ClientBuilder s3ClientBuilder) {
-			return new CrossRegionS3Client(s3ClientBuilder);
+			return new DefaultCrossRegionS3Client(s3ClientBuilder);
 		}
 
 	}
