+++ spring-cloud-aws-s3-parent/spring-cloud-aws-s3-cross-region-client/src/test/java/io/awspring/cloud/s3/crossregion/DefaultCrossRegionS3ClientTests.java
@@ -32,17 +32,18 @@ import software.amazon.awssdk.http.SdkHttpResponse;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.S3ClientBuilder;
+import software.amazon.awssdk.services.s3.model.CreateBucketRequest;
 import software.amazon.awssdk.services.s3.model.HeadBucketRequest;
 import software.amazon.awssdk.services.s3.model.ListObjectsRequest;
 import software.amazon.awssdk.services.s3.model.S3Exception;
 
 /**
- * Unit tests for {@link CrossRegionS3Client}. Integration testing with Localstack is not possible due to:
+ * Unit tests for {@link DefaultCrossRegionS3Client}. Integration testing with Localstack is not possible due to:
  * https://github.com/localstack/localstack/issues/5748
  *
  * @author Maciej Walkowiak
  */
-class CrossRegionS3ClientTests {
+class DefaultCrossRegionS3ClientTests {
 
 	private final S3ClientBuilder mock = mock(S3ClientBuilder.class);
 
@@ -56,7 +57,7 @@ class CrossRegionS3ClientTests {
 	 */
 	private final Map<Region, S3Client> clients = new HashMap<>();
 
-	private CrossRegionS3Client crossRegionS3Client;
+	private DefaultCrossRegionS3Client crossRegionS3Client;
 
 	@BeforeEach
 	void beforeEach() {
@@ -71,7 +72,7 @@ class CrossRegionS3ClientTests {
 				.thenAnswer(invocationOnMock -> builders.get(invocationOnMock.getArgument(0, Region.class)));
 
 		when(mock.build()).thenReturn(defaultClient);
-		crossRegionS3Client = new CrossRegionS3Client(mock);
+		crossRegionS3Client = new DefaultCrossRegionS3Client(mock);
 	}
 
 	@Test
@@ -133,8 +134,8 @@ class CrossRegionS3ClientTests {
 										.sdkHttpResponse(SdkHttpResponse.builder().statusCode(301).build()).build())
 								.build());
 		assertThatThrownBy(() -> crossRegionS3Client.headBucket(r -> r.bucket("first-bucket")))
-				.isInstanceOf(CrossRegionS3Client.RegionDiscoveryException.class);
-		verify(defaultClient, atLeastOnce()).headBucket(HeadBucketRequest.builder().bucket("first-bucket").build());
+				.isInstanceOf(DefaultCrossRegionS3Client.RegionDiscoveryException.class);
+		verify(defaultClient, times(1)).headBucket(HeadBucketRequest.builder().bucket("first-bucket").build());
 	}
 
 	@Test
@@ -150,28 +151,35 @@ class CrossRegionS3ClientTests {
 	}
 
 	@Test
-	void headObjectUsedWhenHeaderMissing() {
+	void headBucketUsedWhenHeaderMissing() {
 		createBucket("first-bucket", Region.EU_WEST_2);
-		crossRegionS3Client.listObjects(r -> r.bucket("first-bucket"));
+		crossRegionS3Client.createBucket(r -> r.bucket("first-bucket"));
 		verify(defaultClient, times(1)).headBucket(HeadBucketRequest.builder().bucket("first-bucket").build());
-		verify(defaultClient, times(1)).listObjects(ListObjectsRequest.builder().bucket("first-bucket").build());
+		verify(defaultClient, times(1)).createBucket(CreateBucketRequest.builder().bucket("first-bucket").build());
 	}
 
 	private void createBucket(String s, Region region) {
 		when(defaultClient.listObjects(any(Consumer.class))).thenCallRealMethod();
-		when(defaultClient.listObjects(ListObjectsRequest.builder().bucket(s).build()))
+		when(defaultClient.listObjects(ListObjectsRequest.builder().bucket(s).build())).thenThrow(S3Exception.builder()
+				.awsErrorDetails(AwsErrorDetails.builder()
+						.sdkHttpResponse(SdkHttpResponse.builder().statusCode(301)
+								.appendHeader(DefaultCrossRegionS3Client.BUCKET_REDIRECT_HEADER, region.id()).build())
+						.build())
+				.build());
+		when(defaultClient.headBucket(any(Consumer.class))).thenCallRealMethod();
+		when(defaultClient.headBucket(HeadBucketRequest.builder().bucket(s).build())).thenThrow(S3Exception.builder()
+				.awsErrorDetails(AwsErrorDetails.builder()
+						.sdkHttpResponse(SdkHttpResponse.builder().statusCode(301)
+								.appendHeader(DefaultCrossRegionS3Client.BUCKET_REDIRECT_HEADER, region.id()).build())
+						.build())
+				.build());
+		when(defaultClient.createBucket(any(Consumer.class))).thenCallRealMethod();
+		when(defaultClient.createBucket(CreateBucketRequest.builder().bucket(s).build()))
 				.thenThrow(
 						S3Exception.builder()
 								.awsErrorDetails(AwsErrorDetails.builder()
 										.sdkHttpResponse(SdkHttpResponse.builder().statusCode(301).build()).build())
 								.build());
-		when(defaultClient.headBucket(any(Consumer.class))).thenCallRealMethod();
-		when(defaultClient.headBucket(HeadBucketRequest.builder().bucket(s).build())).thenThrow(S3Exception.builder()
-				.awsErrorDetails(AwsErrorDetails.builder()
-						.sdkHttpResponse(SdkHttpResponse.builder().statusCode(301)
-								.appendHeader(CrossRegionS3Client.BUCKET_REDIRECT_HEADER, region.id()).build())
-						.build())
-				.build());
 	}
 
 }
